Final version added
In this part of the assignment, we will implement an image stitcher that uses image warping
and homographies to automatically create an image mosaic. We will focus on the case where we have two input images that should form the mosaic, where we warp one image into the plane
of the second image and display the combined views. This problem will give some practice manipulating homogeneous coordinates, computing homography matrices, and performing im- age warps. For simplicity, we will specify corresponding pairs of points manually using mouse clicks.